Consolidate Value::coerce_* methods into value/coerce.rs
`value.rs` had 17 `coerce_to_*` methods on `Value` (plus the static
`coerce_f64_to_int` and the private `coerce_to_rstring_inner` helper)
scattered across ~450 lines. Extract them all into a new
`value/coerce.rs` submodule so the coercion surface lives in one
place.

While moving, also lift `coerce_to_char` out of
`executor/format.rs` — it was a free function `(vm, globals, val)`
purely because of where it was defined, not because it needed to be.
It is now `Value::coerce_to_char(&self, vm, globals)`, called as
`val.coerce_to_char(self, globals)` from the `%c` arm of
`format_by_args`.

No behaviour changes. `cargo test -p monoruby --lib` reports the
same 1548 passed / 130 failed split as the parent commit.
